refactor: densify world connectivity and player-view tests for wave 5 slice E (#4125)
Extract shared blockade/connectivity scenario builders, split near-cap player-view
and province-lookup traversal suites, and clear the test file-size grandfather allowlist.
